+++ src/docs/Docs.vue
@@ -61,6 +61,7 @@
             <z-menu-item :link="{ name: 'docsMenu' }" name="Menu" activeClass="active" />
             <z-menu-item :link="{ name: 'docsModal' }" name="Modals" activeClass="active" />
             <z-menu-item :link="{ name: 'docsNav' }" name="Nav" activeClass="active" />
+            <z-menu-item :link="{ name: 'docsPagination' }" name="Pagination" activeClass="active" />
             <z-menu-item :link="{ name: 'docsStep' }" name="Steps" activeClass="active" />
             <z-menu-item :link="{ name: 'docsTabs' }" name="Tabs" activeClass="active" />
             <z-menu-item :link="{ name: 'docsTile' }" name="Tile" activeClass="active" />