+++ CHANGELOG.md
@@ -24,6 +24,7 @@ All notable changes to this project will be documented in this file.
 * New component [Drawer](https://maclisowski.github.io/zutre/#/components/drawer)
 * New component [Accordion](https://maclisowski.github.io/zutre/#/components/accordions)
 * New component [Tab](https://maclisowski.github.io/zutre/#/components/tabs)
+* New component [Pagination](https://maclisowski.github.io/zutre/#/components/pagination)
 * General cleanups in the code
 * New property in Avatar - `{Boolean} block`, when set to `true`, css properties will be added: `display: block` and auto left/right margins plus centered position inside the block  
 * Added tooltip suppeort: `tooltip` and `tooltipPosition = left/right/bottom/top` for components: Link, Button